+++ src/ts/editor/processKeydown.ts
@@ -7,6 +7,8 @@ import {getSelectPosition} from "./getSelectPosition";
 import {insertText} from "./insertText";
 
 export const processKeydown = (vditor: IVditor, event: KeyboardEvent) => {
+    vditor.undo.recordFirstPosition(vditor);
+
     const editorElement = vditor.editor.element;
     const position = getSelectPosition(editorElement);
     const text = getMarkdown(vditor);